Update w/ PR review comments
