+++ disease/etl/base.py
@@ -17,9 +17,16 @@ class Base(ABC):
         """Public-facing method to begin ETL procedures on given data."""
         raise NotImplementedError
 
-    @abstractmethod
-    def _extract_data(self, *args, **kwargs):
-        raise NotImplementedError
+    def _extract_data(self):
+        """Get source file from data directory."""
+        self._data_path.mkdir(exist_ok=True, parents=True)
+        src_name = type(self).__name__.lower()
+        dir_files = [f for f in self._data_path.iterdir()
+                     if f.name.startswith(src_name)]
+        if len(dir_files) == 0:
+            self._download_data()
+            dir_files = list(self._data_path.iterdir())
+        self._data_file = sorted(dir_files, reverse=True)[0]
 
     @abstractmethod
     def _transform_data(self, *args, **kwargs):
